src/requirements.txt
pika
python-dotenv
mysql-connector
sqlalchemy
pymysql
alembic
rabbitmq-admin